requirements.txt
fastapi
Jinja2
python-dotenv
python-multipart
requests
uvicorn